+++ R/optimise-utils.R
@@ -23,3 +23,17 @@ expr_uses_var <- function(x, vars) {
 
   any(purrr::map_lgl(as.list(x[-1]), expr_uses_var, vars))
 }
+
+any_expr_uses_sql <- function(dots) {
+  purrr::some(dots, expr_uses_sql)
+}
+
+expr_uses_sql <- function(x) {
+  if (is.sql(x)) return(TRUE)
+  if (is_call(x, "sql")) return(TRUE)
+  if (is_quosure(x)) return(expr_uses_sql(quo_get_expr(x)))
+
+  if (!is_call(x)) return(FALSE)
+
+  any(purrr::map_lgl(as.list(x[-1]), expr_uses_sql))
+}