+++ R/translate-sql-conditional.R
@@ -53,21 +53,23 @@ sql_case_match_clause <- function(f, x, con) {
   if (is_call(f_query, "c")) {
     # need to handle `c(...)` specially because on `expr(c(1, y))` it returns
     # `sql('1', '`y`')`
-    f_query <- translate_sql_(call_args(f_query), con)
-  } else if (is_call(f_query) || is_symbol(f_query)) {
-    f_query <- translate_sql(!!f_query, con = con)
+    f_query <- call_args(f_query)
+  } else {
+    if (!is_vector(f_query) || length(f_query) == 1) {
+      f_query <- list(f_query)
+    }
   }
 
-  # NA need to be translated to `IS NULL` instead of `IN (NULL)`
-  # due to the preceeding translation it might be NA or the string NULL
-  missing_loc <- is.na(f_query) | f_query == "NULL"
+  f_query <- purrr::map_if(f_query, ~ !is_vector(.x), ~ translate_sql(!!.x, con = con))
+  missing_loc <- purrr::map_lgl(f_query, ~ is.null(.x) || is.na(.x))
+
   f_query <- vctrs::vec_slice(f_query, !missing_loc)
   has_na <- any(missing_loc)
 
   query <- NULL
   if (!is_empty(f_query)) {
-    values <- sql_vector(f_query, parens = TRUE, collapse = ", ", con = con)
-    query <- translate_sql(!!x %in% !!values)
+    f_query <- escape(f_query, con = con, parens = TRUE, collapse = ", ")
+    query <- translate_sql(!!x %in% !!f_query)
   }
 
   if (has_na) {